Add missing uint64_t cast, improve avr code.
This fixes a bug where the left shift loop was only int-sized, i.e.
16-bit on avr, potnetially causing bad things to happen.  It also
generated terribly inefficient avr code where the result of the loop was
then sign-extended, and'd and or'd byte by byte, etc.  Now avrgcc uses
__lshrdi3 which is more space-efficient and possibly more cpu-efficient,
although still quite subotimal.  Eventually we should get rid of all
64-bit operations.

+++ Sensorino/Sensorino.cpp
@@ -256,7 +256,7 @@ static void SensorinoGPIOISR(int port, uint8_t new_val) {
     for (volatile uint8_t *gpio = &pcint_to_gpio[port << 3]; diff;
             diff >>= 1, gpio++)
         if ((diff & 1) && (uint8_t) ~*gpio) {
-            if (gpio_obj_mask & (1 << *gpio)) {
+            if (gpio_obj_mask & ((uint64_t) 1 << *gpio)) {
                 GenIntrCallback *cb = (GenIntrCallback *) gpio_handler[*gpio];
                 cb->call(*gpio);
             } else {
